requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.1
django-autoslug==1.9.8
django-embed-video==1.4.5
django-jazzmin==2.6.0
django-tinymce==3.5.0
idna==3.4
mysqlclient==2.1.1
Pillow==9.2.0
requests==2.28.1
sqlparse==0.4.3
tzdata==2022.4
urllib3==1.26.12
whitenoise==6.2.0
